RavenDB-22217 - Create AbstractRaftIndexNotifications
