+++ test/SlowTests/Issues/RavenDB-22217.cs
@@ -70,7 +70,7 @@ public class RavenDB_22217: RavenTestBase
                 var delayTask = Task.Delay(200);
                 var task = await Task.WhenAny(saveIndexTask, delayTask);
 
-                WaitForUserToContinueTheTest(store, false);
+                // WaitForUserToContinueTheTest(store, false);
 
                 Assert.Equal(delayTask, task);
